+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ This document records the main changes to the ``sdssdb`` code.
 * Clear metadata on reconnect.
 * Add ``lite`` materialized views to ``catalogdb``.
 * Add ``catalogdb.gaia_assas_sn_cepheids`` table.
+* Move deprecated tables to ``deprecated`` schema in ``sdss5db``.
 
 * :release:`0.4.3 <2020-06-05>`
 * Add schema for ``gaia_dr2_ruwe``.